requirements.txt
gymnasium==0.29.1
matplotlib==3.7.3
numpy==1.24.4
PyYAML==6.0.1
PyYAML==6.0.2
